uncleaned - turned map to strings and store in db

+++ src/main/java/com/mysql/example/demo/responses/mobile/PlayerStatsNBADotComClientResponse.java
@@ -13,7 +13,7 @@ public class PlayerStatsNBADotComClientResponse {
     public double rebounds;
 
     @JsonProperty("assists")
-    public double assists;
+    public Integer assists;
 
     @JsonProperty("steals")
     public double steals;
